# Improve strictness and efficiency of cereal benchmark