--- a/haskell-cereal/mp4.cabal
+++ b/haskell-cereal/mp4.cabal
@@ -20,7 +20,7 @@ executable mp4
   main-is:             Main.hs
   -- other-modules:       
   -- other-extensions:    
-  build-depends:       criterion, cereal, bytestring, base >=4.7 && <5
+  build-depends:       criterion, cereal, bytestring, vector, base >=4.7 && <5
   hs-source-dirs:      src
   default-language:    Haskell2010
   ghc-options:         -O2